First pass at refactoring pdfjs
